Core update: remove TimeseriesEvent
Also rework event resources
