+++ bemserver_api/resources/__init__.py
@@ -8,6 +8,12 @@ from . import timeseries_groups
 from . import timeseries_groups_by_users
 from . import timeseries_groups_by_campaigns
 from . import timeseries_data
+from . import event_states
+from . import event_levels
+from . import event_categories
+from . import event_channels
+from . import event_channels_by_users
+from . import event_channels_by_campaigns
 from . import events
 
 
@@ -20,6 +26,12 @@ MODULES = (
     timeseries_groups_by_users,
     timeseries_groups_by_campaigns,
     timeseries_data,
+    event_states,
+    event_levels,
+    event_categories,
+    event_channels,
+    event_channels_by_users,
+    event_channels_by_campaigns,
     events,
 )
