+++ bemserver_api/resources/event_categories/__init__.py
