+++ requirements/install.txt
@@ -10,7 +10,7 @@ argon2-cffi==21.3.0
     # via bemserver-core
 argon2-cffi-bindings==21.2.0
     # via argon2-cffi
-bemserver-core @ https://github.com/BEMServer/bemserver-core/archive/bcea5ab.tar.gz
+bemserver-core @ https://github.com/BEMServer/bemserver-core/archive/d88ec8a.tar.gz
     # via bemserver-api (setup.py)
 cffi==1.15.0
     # via
