+++ setup.py
@@ -39,7 +39,7 @@ setup(
         (
             # https://github.com/jazzband/pip-tools/issues/1359
             "bemserver-core @ "
-            "https://github.com/BEMServer/bemserver-core/archive/bcea5ab.tar.gz"
+            "https://github.com/BEMServer/bemserver-core/archive/d88ec8a.tar.gz"
         ),
     ],
     packages=find_packages(exclude=["tests*"]),
