+++ tests/resources/test_events.py
@@ -1,4 +1,6 @@
 """Events tests"""
+import datetime as dt
+
 import pytest
 
 from tests.common import AuthHeader
@@ -6,10 +8,11 @@ from tests.common import AuthHeader
 
 DUMMY_ID = "69"
 
+EVENT_STATES_URL = "/event_states/"
+EVENT_LEVELS_URL = "/event_levels/"
+EVENT_CATEGORIES_URL = "/event_categories/"
+EVENT_CHANNELS_URL = "/event_channels/"
 EVENTS_URL = "/events/"
-EVENT_CHANNELS_URL = f"{EVENTS_URL}channels/"
-EVENT_CHANNELS_BY_USERS_URL = f"{EVENTS_URL}channelsbyusers/"
-EVENT_CHANNELS_BY_CAMPAIGNS_URL = f"{EVENTS_URL}channelsbycampaigns/"
 
 
 class TestEventStatesApi:
@@ -24,7 +27,7 @@ class TestEventStatesApi:
         client = app.test_client()
 
         with AuthHeader(creds):
-            ret = client.get(f"{EVENTS_URL}states")
+            ret = client.get(EVENT_STATES_URL)
             assert ret.status_code == 200
             assert len(ret.json) == 3
             for x in ret.json:
@@ -38,7 +41,7 @@ class TestEventStatesApi:
 
         client = app.test_client()
 
-        ret = client.get(f"{EVENTS_URL}states")
+        ret = client.get(EVENT_STATES_URL)
         assert ret.status_code == 401
 
 
@@ -56,7 +59,7 @@ class TestEventLevelsApi:
         with AuthHeader(creds):
 
             # GET level list
-            ret = client.get(f"{EVENTS_URL}levels")
+            ret = client.get(EVENT_LEVELS_URL)
             assert ret.status_code == 200
             assert len(ret.json) == 4
             for x in ret.json:
@@ -71,7 +74,7 @@ class TestEventLevelsApi:
 
         client = app.test_client()
 
-        ret = client.get(f"{EVENTS_URL}levels")
+        ret = client.get(EVENT_LEVELS_URL)
         assert ret.status_code == 401
 
 
@@ -89,7 +92,7 @@ class TestEventCategoriesApi:
         with AuthHeader(creds):
 
             # GET category list
-            ret = client.get(f"{EVENTS_URL}categories")
+            ret = client.get(EVENT_CATEGORIES_URL)
             assert ret.status_code == 200
             assert len(ret.json) > 0
 
@@ -97,236 +100,18 @@ class TestEventCategoriesApi:
 
         client = app.test_client()
 
-        ret = client.get(f"{EVENTS_URL}categories")
+        ret = client.get(EVENT_CATEGORIES_URL)
         assert ret.status_code == 401
 
 
-class TestEventChannelsApi:
-    def test_event_channels_api(self, app, users):
-
-        creds = users["Chuck"]["creds"]
-
-        client = app.test_client()
-
-        with AuthHeader(creds):
-
-            # GET list
-            ret = client.get(EVENT_CHANNELS_URL)
-            assert ret.status_code == 200
-            assert ret.json == []
-
-            # POST
-            event_channel_1 = {
-                "name": "Event channel 1",
-            }
-            ret = client.post(EVENT_CHANNELS_URL, json=event_channel_1)
-            assert ret.status_code == 201
-            ret_val = ret.json
-            event_channel_1_id = ret_val.pop("id")
-            event_channel_1_etag = ret.headers["ETag"]
-            assert ret_val == event_channel_1
-
-            # GET list
-            ret = client.get(EVENT_CHANNELS_URL)
-            assert ret.status_code == 200
-            ret_val = ret.json
-            assert len(ret_val) == 1
-            assert ret_val[0]["id"] == event_channel_1_id
-
-            # GET by id
-            ret = client.get(f"{EVENT_CHANNELS_URL}{event_channel_1_id}")
-            assert ret.status_code == 200
-            assert ret.headers["ETag"] == event_channel_1_etag
-            ret_val = ret.json
-            ret_val.pop("id")
-            assert ret_val == event_channel_1
-
-            # PUT
-            event_channel_1["name"] = "Great event channel 1"
-            ret = client.put(
-                f"{EVENT_CHANNELS_URL}{event_channel_1_id}",
-                json=event_channel_1,
-                headers={"If-Match": event_channel_1_etag},
-            )
-            assert ret.status_code == 200
-            ret_val = ret.json
-            ret_val.pop("id")
-            event_channel_1_etag = ret.headers["ETag"]
-            assert ret_val == event_channel_1
-
-            # PUT wrong ID -> 404
-            ret = client.put(
-                f"{EVENT_CHANNELS_URL}{DUMMY_ID}",
-                json=event_channel_1,
-                headers={"If-Match": event_channel_1_etag},
-            )
-            assert ret.status_code == 404
-
-            # POST event_channel 2
-            event_channel_2 = {
-                "name": "Event channel 2",
-            }
-            ret = client.post(EVENT_CHANNELS_URL, json=event_channel_2)
-            ret_val = ret.json
-            event_channel_2_id = ret_val.pop("id")
-            event_channel_2_etag = ret.headers["ETag"]
-
-            # GET list
-            ret = client.get(EVENT_CHANNELS_URL)
-            assert ret.status_code == 200
-            ret_val = ret.json
-            assert len(ret_val) == 2
-
-            # GET list with filters
-            ret = client.get(
-                EVENT_CHANNELS_URL, query_string={"name": "Great event channel 1"}
-            )
-            assert ret.status_code == 200
-            ret_val = ret.json
-            assert len(ret_val) == 1
-            assert ret_val[0]["id"] == event_channel_1_id
-
-            # DELETE wrong ID -> 404
-            ret = client.delete(
-                f"{EVENT_CHANNELS_URL}{DUMMY_ID}",
-                headers={"If-Match": event_channel_1_etag},
-            )
-            assert ret.status_code == 404
-
-            # DELETE
-            ret = client.delete(
-                f"{EVENT_CHANNELS_URL}{event_channel_1_id}",
-                headers={"If-Match": event_channel_1_etag},
-            )
-            assert ret.status_code == 204
-            ret = client.delete(
-                f"{EVENT_CHANNELS_URL}{event_channel_2_id}",
-                headers={"If-Match": event_channel_2_etag},
-            )
-            assert ret.status_code == 204
-
-            # GET list
-            ret = client.get(EVENT_CHANNELS_URL)
-            assert ret.status_code == 200
-            ret_val = ret.json
-            assert len(ret_val) == 0
-
-            # GET by id -> 404
-            ret = client.get(f"{EVENT_CHANNELS_URL}{event_channel_1_id}")
-            assert ret.status_code == 404
-
-    @pytest.mark.usefixtures("event_channels_by_users")
-    def test_event_channels_as_user_api(self, app, users, event_channels):
-
-        user_creds = users["Active"]["creds"]
-        event_channel_1_id, event_channel_2_id = event_channels
-
-        client = app.test_client()
-
-        with AuthHeader(user_creds):
-
-            # GET list
-            ret = client.get(EVENT_CHANNELS_URL)
-            assert ret.status_code == 200
-            ret_val = ret.json
-            assert len(ret_val) == 1
-            event_channel_1 = ret_val[0]
-            assert event_channel_1.pop("id") == event_channel_1_id
-
-            # GET list with filters
-            ret = client.get(
-                EVENT_CHANNELS_URL, query_string={"name": "Event channel 1"}
-            )
-            assert ret.status_code == 200
-            ret_val = ret.json
-            assert len(ret_val) == 1
-            assert ret_val[0]["id"] == event_channel_1_id
-            ret = client.get(
-                EVENT_CHANNELS_URL, query_string={"name": "Event channel 2"}
-            )
-            assert ret.status_code == 200
-            assert not ret.json
-
-            # POST
-            event_channel_3 = {
-                "name": "Event channel 3",
-            }
-            ret = client.post(EVENT_CHANNELS_URL, json=event_channel_3)
-            assert ret.status_code == 403
-
-            # GET by id
-            ret = client.get(f"{EVENT_CHANNELS_URL}{event_channel_1_id}")
-            assert ret.status_code == 200
-            event_channel_1_etag = ret.headers["ETag"]
-
-            ret = client.get(f"{EVENT_CHANNELS_URL}{event_channel_2_id}")
-            assert ret.status_code == 403
-
-            # PUT
-            event_channel_1["name"] = "Great event channel 1"
-            ret = client.put(
-                f"{EVENT_CHANNELS_URL}{event_channel_1_id}",
-                json=event_channel_1,
-                headers={"If-Match": event_channel_1_etag},
-            )
-            assert ret.status_code == 403
-
-            # DELETE
-            ret = client.delete(
-                f"{EVENT_CHANNELS_URL}{event_channel_1_id}",
-                headers={"If-Match": event_channel_1_etag},
-            )
-            assert ret.status_code == 403
-
-    def test_event_channels_as_anonym_api(self, app, event_channels):
-
-        event_channel_1_id, _ = event_channels
-
-        client = app.test_client()
-
-        # GET list
-        ret = client.get(EVENT_CHANNELS_URL)
-        assert ret.status_code == 401
-
-        # POST
-        event_channel_3 = {
-            "name": "Event channel 3",
-        }
-        ret = client.post(EVENT_CHANNELS_URL, json=event_channel_3)
-        assert ret.status_code == 401
-
-        # GET by id
-        ret = client.get(f"{EVENT_CHANNELS_URL}{event_channel_1_id}")
-        assert ret.status_code == 401
-
-        # PUT
-        event_channel_1 = {
-            "name": "Super Event channel 1",
-        }
-        ret = client.put(
-            f"{EVENT_CHANNELS_URL}{event_channel_1_id}",
-            json=event_channel_1,
-            headers={"If-Match": "Dummy-ETag"},
-        )
-        # ETag is wrong but we get rejected before ETag check anyway
-        assert ret.status_code == 401
-
-        # DELETE
-        ret = client.delete(
-            f"{EVENT_CHANNELS_URL}{event_channel_1_id}",
-            headers={"If-Match": "Dummy-Etag"},
-        )
-        # ETag is wrong but we get rejected before ETag check anyway
-        assert ret.status_code == 401
-
-
-class TestEventChannelsByCampaignsApi:
-    def test_event_channels_by_campaigns_api(
-        self, app, users, event_channels, campaigns
-    ):
+class TestEventsApi:
+    def test_events_api(self, app, users, event_channels):
         ec_1_id = event_channels[0]
         ec_2_id = event_channels[1]
-        campaign_1_id, campaign_2_id = campaigns
+        event_channel_1_id = event_channels[0]
+        event_channel_2_id = event_channels[1]
+        c1_st = dt.datetime(2020, 1, 1, tzinfo=dt.timezone.utc).isoformat()
+        c2_et = dt.datetime(2021, 1, 1, tzinfo=dt.timezone.utc).isoformat()
 
         creds = users["Chuck"]["creds"]
 
@@ -335,124 +120,123 @@ class TestEventChannelsByCampaignsApi:
         with AuthHeader(creds):
 
             # GET list
-            ret = client.get(f"{EVENTS_URL}channelsbycampaigns/")
+            ret = client.get(EVENTS_URL)
             assert ret.status_code == 200
             assert ret.json == []
 
             # POST
-            ecbc_1 = {
-                "campaign_id": campaign_1_id,
-                "event_channel_id": ec_1_id,
+            tse_1 = {
+                "channel_id": event_channel_1_id,
+                "timestamp": c1_st,
+                "source": "Event source",
+                "category": "observation_missing",
+                "level": "INFO",
+                "state": "NEW",
             }
-            ret = client.post(EVENT_CHANNELS_BY_CAMPAIGNS_URL, json=ecbc_1)
+            ret = client.post(EVENTS_URL, json=tse_1)
             assert ret.status_code == 201
             ret_val = ret.json
-            ecbc_1_id = ret_val.pop("id")
-            ecbc_1_etag = ret.headers["ETag"]
+            tse_1_id = ret_val.pop("id")
+            tse_1_etag = ret.headers["ETag"]
 
-            # POST violating unique constraint
-            ret = client.post(EVENT_CHANNELS_BY_CAMPAIGNS_URL, json=ecbc_1)
+            # POST violating fkey constraint
+            tse_1["level"] = "Dummy Level"
+            ret = client.post(EVENTS_URL, json=tse_1)
             assert ret.status_code == 409
 
             # GET list
-            ret = client.get(EVENT_CHANNELS_BY_CAMPAIGNS_URL)
+            ret = client.get(EVENTS_URL)
             assert ret.status_code == 200
             ret_val = ret.json
             assert len(ret_val) == 1
-            assert ret_val[0]["id"] == ecbc_1_id
+            assert ret_val[0]["id"] == tse_1_id
 
             # GET by id
-            ret = client.get(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_1_id}")
+            ret = client.get(f"{EVENTS_URL}{tse_1_id}")
             assert ret.status_code == 200
-            assert ret.headers["ETag"] == ecbc_1_etag
+            assert ret.headers["ETag"] == tse_1_etag
 
             # POST
-            ecbc_2 = {
-                "campaign_id": campaign_2_id,
-                "event_channel_id": ec_2_id,
+            tse_2 = {
+                "channel_id": event_channel_2_id,
+                "timestamp": c2_et,
+                "source": "Another event source",
+                "category": "observation_missing",
+                "level": "WARNING",
+                "state": "NEW",
             }
-            ret = client.post(EVENT_CHANNELS_BY_CAMPAIGNS_URL, json=ecbc_2)
+            ret = client.post(EVENTS_URL, json=tse_2)
             assert ret.status_code == 201
             ret_val = ret.json
-            ecbc_2_id = ret_val.pop("id")
+            tse_2_id = ret_val.pop("id")
+            tse_2_etag = ret.headers["ETag"]
 
             # GET list (filtered)
-            ret = client.get(
-                EVENT_CHANNELS_BY_CAMPAIGNS_URL,
-                query_string={"event_channel_id": ec_1_id},
-            )
+            ret = client.get(EVENTS_URL, query_string={"channel_id": ec_1_id})
             assert ret.status_code == 200
             ret_val = ret.json
             assert len(ret_val) == 1
-            assert ret_val[0]["id"] == ecbc_1_id
-            assert ret_val[0]["event_channel_id"] == ec_1_id
-            assert ret_val[0]["campaign_id"] == campaign_1_id
-            ret = client.get(
-                EVENT_CHANNELS_BY_CAMPAIGNS_URL,
-                query_string={"campaign_id": campaign_2_id},
-            )
+            assert ret_val[0]["id"] == tse_1_id
+            ret = client.get(EVENTS_URL, query_string={"level": "WARNING"})
             assert ret.status_code == 200
             ret_val = ret.json
             assert len(ret_val) == 1
-            assert ret_val[0]["id"] == ecbc_2_id
-            assert ret_val[0]["event_channel_id"] == ec_2_id
-            assert ret_val[0]["campaign_id"] == ec_2_id
+            assert ret_val[0]["id"] == tse_2_id
             ret = client.get(
-                EVENT_CHANNELS_BY_CAMPAIGNS_URL,
+                EVENTS_URL,
                 query_string={
-                    "event_channel_id": ec_1_id,
-                    "campaign_id": campaign_2_id,
+                    "channel_id": ec_2_id,
+                    "level": "INFO",
                 },
             )
             assert ret.status_code == 200
             ret_val = ret.json
             assert len(ret_val) == 0
 
-            # GET channels list filtered by campaign
-            ret = client.get(
-                EVENT_CHANNELS_URL, query_string={"campaign_id": campaign_1_id}
-            )
-            assert ret.status_code == 200
-            ret_val = ret.json
-            assert len(ret_val) == 1
-            assert ret_val[0]["id"] == ec_1_id
-
             # DELETE wrong ID -> 404
-            ret = client.delete(f"{EVENTS_URL}channels_by_campaigns/{DUMMY_ID}")
+            # ETag is wrong but we get rejected before ETag check anyway
+            ret = client.delete(
+                f"{EVENTS_URL}{DUMMY_ID}",
+                headers={"If-Match": "Dummy-ETag"},
+            )
             assert ret.status_code == 404
 
             # DELETE channel violating fkey constraint
             ret = client.get(f"{EVENT_CHANNELS_URL}{ec_1_id}")
             ec_1_etag = ret.headers["ETag"]
             ret = client.delete(
-                f"{EVENT_CHANNELS_URL}{ec_1_id}", headers={"If-Match": ec_1_etag}
+                f"{EVENT_CHANNELS_URL}{ec_1_id}",
+                headers={"If-Match": ec_1_etag},
             )
             assert ret.status_code == 409
 
             # DELETE
-            ret = client.delete(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_1_id}")
+            ret = client.delete(
+                f"{EVENTS_URL}{tse_1_id}",
+                headers={"If-Match": tse_1_etag},
+            )
             assert ret.status_code == 204
-            ret = client.delete(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_2_id}")
+            ret = client.delete(
+                f"{EVENTS_URL}{tse_2_id}",
+                headers={"If-Match": tse_2_etag},
+            )
             assert ret.status_code == 204
 
             # GET list
-            ret = client.get(EVENT_CHANNELS_BY_CAMPAIGNS_URL)
+            ret = client.get(EVENTS_URL)
             assert ret.status_code == 200
             ret_val = ret.json
             assert len(ret_val) == 0
 
             # GET by id -> 404
-            ret = client.get(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_1_id}")
+            ret = client.get(f"{EVENTS_URL}{tse_1_id}")
             assert ret.status_code == 404
 
-    @pytest.mark.usefixtures("users_by_campaigns")
-    def test_event_channel_by_users_as_user_api(
-        self, app, users, event_channels, event_channels_by_users
-    ):
-        ec_1_id = event_channels[0]
-        ecbu_1_id = event_channels_by_users[0]
-        ecbu_2_id = event_channels_by_users[1]
-        user_1_id = users["Active"]["id"]
+    @pytest.mark.usefixtures("event_channels_by_users")
+    def test_events_as_user_api(self, app, users, event_channels, campaigns, events):
+        tse_1_id = events[0]
+        event_channel_1_id = event_channels[0]
+        c1_st = dt.datetime(2021, 1, 1, tzinfo=dt.timezone.utc).isoformat()
 
         creds = users["Active"]["creds"]
 
@@ -461,114 +245,65 @@ class TestEventChannelsByCampaignsApi:
         with AuthHeader(creds):
 
             # GET list
-            ret = client.get(EVENT_CHANNELS_BY_USERS_URL)
+            ret = client.get(EVENTS_URL)
             assert ret.status_code == 200
             assert len(ret.json) == 1
 
             # POST
-            ecbu = {"user_id": user_1_id, "event_channel_id": ec_1_id}
-            ret = client.post(EVENT_CHANNELS_BY_USERS_URL, json=ecbu)
-            assert ret.status_code == 403
+            tse = {
+                "channel_id": event_channel_1_id,
+                "timestamp": c1_st,
+                "source": "Event source",
+                "category": "observation_missing",
+                "level": "INFO",
+                "state": "NEW",
+            }
+            ret = client.post(EVENTS_URL, json=tse)
+            assert ret.status_code == 201
 
             # GET by id
-            ret = client.get(f"{EVENT_CHANNELS_BY_USERS_URL}{ecbu_1_id}")
+            ret = client.get(f"{EVENTS_URL}{tse_1_id}")
             assert ret.status_code == 200
-
-            # GET by id
-            ret = client.get(f"{EVENT_CHANNELS_BY_USERS_URL}{ecbu_2_id}")
-            assert ret.status_code == 403
+            tse_1_etag = ret.headers["ETag"]
 
             # DELETE
-            ret = client.delete(f"{EVENT_CHANNELS_BY_USERS_URL}{ecbu_1_id}")
-            assert ret.status_code == 403
+            ret = client.delete(
+                f"{EVENTS_URL}{tse_1_id}",
+                headers={"If-Match": tse_1_etag},
+            )
+            assert ret.status_code == 204
 
-    @pytest.mark.usefixtures("users_by_campaigns")
-    def test_event_channel_by_users_as_anonym_api(
-        self, app, users, event_channels, event_channels_by_users
-    ):
-        ec_1_id = event_channels[0]
-        ecbu_1_id = event_channels_by_users[0]
-        user_1_id = users["Active"]["id"]
+    def test_events_as_anonym_api(self, app, event_channels, events):
+        tse_1_id = events[0]
+        event_channel_1_id = event_channels[0]
+        c1_st = dt.datetime(2021, 1, 1, tzinfo=dt.timezone.utc).isoformat()
 
         client = app.test_client()
 
         # GET list
-        ret = client.get(EVENT_CHANNELS_BY_USERS_URL)
+        ret = client.get(EVENTS_URL)
         assert ret.status_code == 401
 
         # POST
-        ecbu = {"user_id": user_1_id, "event_channel_id": ec_1_id}
-        ret = client.post(EVENT_CHANNELS_BY_USERS_URL, json=ecbu)
+        tse = {
+            "channel_id": event_channel_1_id,
+            "timestamp": c1_st,
+            "source": "Event source",
+            "category": "observation_missing",
+            "level": "INFO",
+            "state": "NEW",
+        }
+        ret = client.post(EVENTS_URL, json=tse)
         assert ret.status_code == 401
 
         # GET by id
-        ret = client.get(f"{EVENT_CHANNELS_BY_USERS_URL}{ecbu_1_id}")
+        ret = client.get(f"{EVENTS_URL}{tse_1_id}")
         assert ret.status_code == 401
 
         # DELETE
-        ret = client.delete(f"{EVENT_CHANNELS_BY_USERS_URL}{ecbu_1_id}")
-        assert ret.status_code == 401
-
-    @pytest.mark.usefixtures("users_by_campaigns")
-    def test_event_channel_by_campaigns_as_user_api(
-        self, app, users, event_channels, campaigns, event_channels_by_campaigns
-    ):
-        ec_1_id = event_channels[0]
-        ecbc_1_id = event_channels_by_campaigns[0]
-        ecbc_2_id = event_channels_by_campaigns[1]
-        campaign_1_id = campaigns[0]
-
-        creds = users["Active"]["creds"]
-
-        client = app.test_client()
-
-        with AuthHeader(creds):
-
-            # GET list
-            ret = client.get(EVENT_CHANNELS_BY_CAMPAIGNS_URL)
-            assert ret.status_code == 200
-            assert len(ret.json) == 1
-
-            # POST
-            ecbc = {"campaign_id": campaign_1_id, "event_channel_id": ec_1_id}
-            ret = client.post(EVENT_CHANNELS_BY_CAMPAIGNS_URL, json=ecbc)
-            assert ret.status_code == 403
-
-            # GET by id
-            ret = client.get(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_1_id}")
-            assert ret.status_code == 200
-
-            # GET by id
-            ret = client.get(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_2_id}")
-            assert ret.status_code == 403
-
-            # DELETE
-            ret = client.delete(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_1_id}")
-            assert ret.status_code == 403
-
-    @pytest.mark.usefixtures("users_by_campaigns")
-    def test_event_channel_by_campaigns_as_anonym_api(
-        self, app, users, event_channels, campaigns, event_channels_by_campaigns
-    ):
-        ec_1_id = event_channels[0]
-        ecbc_1_id = event_channels_by_campaigns[0]
-        campaign_1_id = campaigns[0]
-
-        client = app.test_client()
-
-        # GET list
-        ret = client.get(EVENT_CHANNELS_BY_CAMPAIGNS_URL)
-        assert ret.status_code == 401
-
-        # POST
-        ecbc = {"campaign_id": campaign_1_id, "event_channel_id": ec_1_id}
-        ret = client.post(EVENT_CHANNELS_BY_CAMPAIGNS_URL, json=ecbc)
-        assert ret.status_code == 401
-
-        # GET by id
-        ret = client.get(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_1_id}")
-        assert ret.status_code == 401
-
-        # DELETE
-        ret = client.delete(f"{EVENT_CHANNELS_BY_CAMPAIGNS_URL}{ecbc_1_id}")
+        # ETag is wrong but we get rejected before ETag check anyway
+        ret = client.delete(
+            f"{EVENTS_URL}{tse_1_id}",
+            headers={"If-Match": "Dummy-ETag"},
+        )
         assert ret.status_code == 401
